fix template setting

+++ src/views/setting/Index.vue
@@ -1,22 +1,22 @@
 <template>
-  <div id="app">
+  <div id="app" ref="laylout">
     <v-app id="inspire">
       <div>
         <v-toolbar flat color="white">
-          <v-spacer></v-spacer>
-          <h3>Thiết lập thông tin công ty</h3>
-          <v-spacer></v-spacer>
+          <h3 style="margin-left:300px">Thiết lập thông tin công ty</h3>
           <v-dialog v-model="dialog" max-width="500px">
+            <v-spacer></v-spacer>
             <v-btn slot="activator" class="mr-5" icon color="primary">
               <v-icon>add</v-icon>
             </v-btn>
+            <v-spacer></v-spacer>
             <v-card>
               <v-card-title>
                 <span class="headline">{{ formTitle }}</span>
               </v-card-title>
               <!-- form edit -->
               <v-card-text id="formSub">
-                <v-container grid-list-md>
+                <v-container grid-list-md class="white scroll-y border-e0-top" :style="{height: dataViewHeight + 'px'}">
                   <v-layout wrap>
                     <!-- name -->
                     <v-flex xs12 sm6 md12>
@@ -64,23 +64,26 @@
           </v-card>
         </v-dialog>
       </v-toolbar>
-      <v-data-table
-      v-if="Array.isArray(settingDetail)"
-      :headers="headers"
-      :items="settingDetail"
-      hide-actions
-      expand
-      class="elevation-1"
-      >
-      <template slot="items" slot-scope="props">
-        <td> {{ props.index + 1 }}</td>
-        <td style="text-transform: capitalize">{{ props.item.name }}</td>
-        <td style="text-transform: capitalize">{{ props.item.value }}</td>
-        <td>{{ props.item.status_txt }}</td>
-        <td id="action"> <v-icon v-if="canAccess('setting.update')" class="mr-6" @click="editItem(props.item,props.item.id)" color="green"> edit</v-icon></td>
-        <td id="action"><v-icon v-if="canAccess('setting.delete')" icon @click="removeConfirm(props.item.id)" color="red"> delete </v-icon></td>
-      </template>
-    </v-data-table>
+      <v-container>
+        <v-data-table
+        v-if="Array.isArray(settingDetail)"
+        :headers="headers"
+        :items="settingDetail"
+        hide-actions
+        expand
+        class="elevation-1"
+        >
+        <template slot="items" slot-scope="props">
+          <td> {{ props.index + 1 }}</td>
+          <td style="text-transform: capitalize">{{ props.item.name }}</td>
+          <td style="text-transform: capitalize">{{ props.item.value }}</td>
+          <td>{{ props.item.status_txt }}</td>
+          <td id="action"> <v-icon v-if="canAccess('setting.update')" class="mr-6" @click="editItem(props.item,props.item.id)" color="green"> edit</v-icon>
+            <v-icon style="margin-left:10px" v-if="canAccess('setting.delete')" icon @click="removeConfirm(props.item.id)" color="red"> delete </v-icon>
+          </td>
+        </template>
+      </v-data-table>
+    </v-container>
     <dialog-confirm v-model="dialogDelete" @input="remove()" />
   </div>
 </v-app>
@@ -94,6 +97,7 @@ export default{
     DialogConfirm
   },
   data: () => ({
+    dataViewHeight: 0,
     idSetting: null,
     dialogDelete: false,
     dialog: false,
@@ -102,8 +106,7 @@ export default{
     { text: 'Tên thông tin', sortable: false },
     { text: 'Giá trị', sortable: false },
     { text: 'Trạng Thái', sortable: false },
-    { text: 'Sửa', sortable: false },
-    { text: 'Xóa', sortable: false }
+    { text: 'Hành động', sortable: false }
     ],
     editedIndex: -1,
     setting: {
@@ -120,15 +123,17 @@ export default{
     ...mapGetters('Setting', ['settingDetail']),
     ...mapGetters(['isFetchingApi'])
   },
-
   watch: {
     dialog (val) {
       val || this.close()
     }
   },
-
   created () {
     this.FetchSetting()
+    this.setMiniDrawer(true)
+  },
+  mounted () {
+    this.dataViewHeight = this.$refs.laylout.clientHeight - 48
   },
   methods: {
     ...mapActions(['setMiniDrawer']),
@@ -146,25 +151,27 @@ export default{
       this.dialogDelete = true
     },
     remove (confirm) {
-      this.deleteSetting({
-        id: this.idSetting,
-        cb: (response) => {
-          this.$store.dispatch('showNotify', {
-            text: this.$t('alert.success'),
-            color: 'success'
-          })
-          this.dialogDelete = false
-          this.FetchSetting()
-        },
-        error: (error) => {
-          if (error.status === 404) {
+      if (confirm) {
+        this.deleteSetting({
+          id: this.idSetting,
+          cb: (response) => {
             this.$store.dispatch('showNotify', {
-              text: this.$t('alert.not-found'),
-              color: 'warning'
+              text: this.$t('alert.success'),
+              color: 'success'
             })
+            this.dialogDelete = false
+            this.FetchSetting()
+          },
+          error: (error) => {
+            if (error.status === 404) {
+              this.$store.dispatch('showNotify', {
+                text: this.$t('alert.not-found'),
+                color: 'warning'
+              })
+            }
           }
-        }
-      })
+        })
+      }
     },
     close () {
       this.dialog = false
@@ -212,22 +219,3 @@ export default{
   }
 }
 </script>
-<style scope>
-p span{
-  padding: 7px;
-  margin-top:-25px;
-  float: right;
-  margin-right:10px;
-}
-h3{
-  clear:both;
-}
-tr td{
-  padding-left:70px !important;
-}
-tr th{
-  padding-left:70px !important;
-  font-size: 16px !important;
-  color: black !important;
-}
-</style>
